backend-py/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
redis-om==0.2.1
langchain==0.1.0
langchain-anthropic==0.1.1
anthropic==0.8.1
sanity==3.0.0
mcp==0.9.0
python-dotenv==1.0.0
httpx==0.25.2
aioredis==2.0.1
python-multipart==0.0.6
email-validator==2.1.0
